+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Change Log
 * Fixed ETA for server-side validation/submission progress bar.
 * Cleaned up messaging for check-submission.
 * Added option to --version to automatically (after prompting) update version to latest.
+* Added metadata_template.py module with goal of checking the users' metadata
+  file with the latest HMS DBMI metadata template and giving a warning if the
+  version appears to be out of date.
 
 
 0.8.0